+++ arch/xtensa/core/irq_manage.c
@@ -5,8 +5,7 @@
 
 #include <zephyr/types.h>
 #include <stdio.h>
-#include <arch/xtensa/xtensa_api.h>
-#include <kernel_arch_data.h>
+#include <arch/xtensa/irq.h>
 #include <sys/__assert.h>
 /*
  * @internal
