+++ arch/xtensa/core/irq_offload.c
@@ -6,7 +6,6 @@
 #include <kernel.h>
 #include <irq_offload.h>
 #include <arch/xtensa/arch.h>
-#include <arch/xtensa/xtensa_api.h>
 
 /*
  * Xtensa core should support software interrupt in order to allow using
