+++ arch/xtensa/core/window_vectors.S
@@ -2,7 +2,7 @@
  * Copyright (c) 2016 Cadence Design Systems, Inc.
  * SPDX-License-Identifier: Apache-2.0
  */
-#include <arch/xtensa/xtensa_rtos.h>
+#include <xtensa/coreasm.h>
 
 /* WINDOW OVERFLOW AND UNDERFLOW EXCEPTION VECTORS AND ALLOCA EXCEPTION
  * HANDLER
