+++ include/arch/xtensa/exc.h
@@ -20,17 +20,6 @@ extern "C" {
 #endif
 
 #ifndef _ASMLANGUAGE
-/**
- * @brief Exception Stack Frame
- *
- * A pointer to an "exception stack frame" (ESF) is passed as an argument
- * to exception handlers registered via nanoCpuExcConnect().
- */
-struct __esf {
-	/* FIXME - not finished yet */
-	sys_define_gpr_with_alias(a1, sp);
-	uint32_t pc;
-};
 
 /* Xtensa uses a variable length stack frame depending on how many
  * register windows are in use.  This isn't a struct type, it just
