+++ include/arch/xtensa/irq.h
@@ -6,10 +6,50 @@
 #ifndef ZEPHYR_INCLUDE_ARCH_XTENSA_XTENSA_IRQ_H_
 #define ZEPHYR_INCLUDE_ARCH_XTENSA_XTENSA_IRQ_H_
 
-#include <arch/xtensa/xtensa_api.h>
 #include <xtensa/xtruntime.h>
 
 #define CONFIG_GEN_IRQ_START_VECTOR 0
+
+/*
+ * Call this function to enable the specified interrupts.
+ *
+ * mask     - Bit mask of interrupts to be enabled.
+ */
+static inline void z_xt_ints_on(unsigned int mask)
+{
+	int val;
+
+	__asm__ volatile("rsr.intenable %0" : "=r"(val));
+	val |= mask;
+	__asm__ volatile("wsr.intenable %0; rsync" : : "r"(val));
+}
+
+
+/*
+ * Call this function to disable the specified interrupts.
+ *
+ * mask     - Bit mask of interrupts to be disabled.
+ */
+static inline void z_xt_ints_off(unsigned int mask)
+{
+	int val;
+
+	__asm__ volatile("rsr.intenable %0" : "=r"(val));
+	val &= ~mask;
+	__asm__ volatile("wsr.intenable %0; rsync" : : "r"(val));
+}
+
+/*
+ * Call this function to set the specified (s/w) interrupt.
+ */
+static inline void z_xt_set_intset(unsigned int arg)
+{
+#if XCHAL_HAVE_INTERRUPTS
+	__asm__ volatile("wsr.intset %0; rsync" : : "r"(arg));
+#else
+	ARG_UNUSED(arg);
+#endif
+}
 
 #ifdef CONFIG_MULTI_LEVEL_INTERRUPTS
