+++ soc/xtensa/intel_adsp/common/soc.c
@@ -5,7 +5,6 @@
  */
 
 #include <device.h>
-#include <arch/xtensa/xtensa_api.h>
 #include <xtensa/xtruntime.h>
 #include <irq_nextlevel.h>
 #include <xtensa/hal.h>
